pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "msr-claw"
version = "0.1.0"
description = "Human-in-the-loop multi-agent CLI for MSR research pipelines"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
dependencies = [
    "langgraph>=1.0",
    "langchain-core>=0.3",
    "langgraph-checkpoint-sqlite>=2.0",
    "typer>=0.12",
    "pyyaml>=6",
    "httpx>=0.27",
    "openai>=1.50",
    "pandas>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "ruff>=0.8",
]
mcp = [
    "mcp>=1.2",
]

[project.scripts]
msr-claw = "msrclaw.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["msrclaw"]

[tool.hatch.build]
include = [
    "msrclaw/**/*.py",
    "docs/**/*.md",
    "MSRCLAW_AGENTS.md",
    "skills/**/*.md",
]

[tool.ruff]
line-length = 100
target-version = "py311"
